pass 5 and pintos -- run alarm-multiple

+++ src/threads/thread.h
@@ -92,7 +92,7 @@ struct thread
 
     /* Shared between thread.c and synch.c. */
     struct list_elem elem;              /* List element. */
-    uint8_t block_ticks_remain;              //calculate how many ticks remain to unblock thread
+    uint16_t block_ticks_remain;              //calculate how many ticks remain to unblock thread
 
 #ifdef USERPROG
     /* Owned by userprog/process.c. */
